requirements.txt
# Requirements for Azure OpenAI Training Data Generator
openai>=1.0.0
azure-identity>=1.12.0
requests>=2.28.0
python-dotenv>=1.0.0
